장바구니 버그 수정
