+++ ios/Mukkaebie/Mukkaebie/StoreDetailViewController.swift
@@ -1,96 +1,158 @@
 //
-//  StoreDetailViewController.swift
-//  MenuLayout
+//  StoreTestViewController.swift
+//  Mukkaebie
 //
-//  Created by woowabrothers on 2017. 8. 8..
-//  Copyright © 2017년 hyeongjong. All rights reserved.
+//  Created by woowabrothers on 2017. 8. 15..
+//  Copyright © 2017년 woowabrothers. All rights reserved.
 //
 
 import UIKit
 
-class StoreDetailViewController: UIViewController, UITabBarDelegate {
+class StoreDetailViewController: UIViewController {
     
-    var mukkaebieRankViewController : MukkaebieRankViewController?
-    var menuViewController: MenuViewController?
-//    var infoViewController: InfoViewController?
-//    var reviewViewController: ReviewViewController?
+    lazy var mukkaebieVC : MukkaebieRankViewController? = {
+        let storyboard = UIStoryboard(name: "MukkaebieRank", bundle: nil)
+        let mukkaebieVC = storyboard.instantiateViewController(withIdentifier: "MukkaebieRank") as? MukkaebieRankViewController
+        
+        return mukkaebieVC
+    }()
     
+    lazy var menuRankVC : MenuViewController? = {
+        let storyboard = UIStoryboard(name: "MenuView", bundle: nil)
+        let menuRankVC = storyboard.instantiateViewController(withIdentifier: "Menu") as? MenuViewController
+        
+        return menuRankVC
+    }()
+    
+    lazy var infoVC : InfoViewController? = {
+        let storyboard = UIStoryboard(name: "Info", bundle: nil)
+        let infoVC = storyboard.instantiateViewController(withIdentifier: "Info") as? InfoViewController
+
+        return infoVC
+    }()
+    
+    lazy var reviewVC : ReviewTableViewController? = {
+        let storyboard = UIStoryboard(name: "Review", bundle: nil)
+        let reviewVC = storyboard.instantiateViewController(withIdentifier: "Review") as? ReviewTableViewController
+        
+        return reviewVC
+    }()
+    
+    var storeId = Int()
     var modelStore : ModelStores?
+    let networkStore = NetworkStore()
+    
+    var priceByMenu = [String:Int]()
+    
     let networkOrder = NetworkOrder()
     var orderList = [ModelOrders]()
-    var priceByMenu = [String:Int]()
+    var orderByUser = [String:Int]()
+    var orderByUserTop3 = [(key: String, value: Int)]()
     var orderByMenu = [String:Int]()
     var orderByMenuSorted = [(key: String, value: Int)]()
     
-    @IBOutlet weak var meetPaymentLabel: UILabel!
-    @IBOutlet weak var directPaymentLabel: UILabel!
+    let indicatorView = Indicator()
+
+    var tabNumber = 0
     
-    @IBOutlet weak var scrollView: UIScrollView!
-    @IBOutlet weak var contentView: UIView!
-    
-    @IBOutlet weak var uppperBarView: UIView!
-    @IBOutlet weak var imageView: UIImageView!
-    @IBOutlet weak var scoreView: UIView!
-    @IBOutlet weak var tabView: UIView!
-    @IBOutlet weak var tabBar: UITabBar!
-    @IBOutlet weak var tabSubView: UIView!
-    
-    @IBOutlet weak var mukkabieTabItem: UITabBarItem!
-    @IBOutlet weak var menuTabItem: UITabBarItem!
-    @IBOutlet weak var infoTabItem: UITabBarItem!
-    @IBOutlet weak var reviewTabItem: UITabBarItem!
-    
+    @IBOutlet weak var tableView: UITableView!
+    @IBOutlet var cartAlertView: UIView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.tabBarController?.tabBar.isHidden = true
-        self.tabBar.delegate = self
         
-        self.automaticallyAdjustsScrollViewInsets = false
-        scrollView.contentInset = UIEdgeInsets.zero
-        scrollView.scrollIndicatorInsets = UIEdgeInsets.zero
-        scrollView.contentOffset = CGPoint(x: 0.0, y: 0.0)
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.allowsSelection = false
+        tableView.rowHeight = UITableViewAutomaticDimension
+        tableView.estimatedRowHeight = 100
         
-        self.navigationItem.title = modelStore?.name
-        
-        meetPaymentLabel.layer.borderWidth = 1
-        meetPaymentLabel.layer.borderColor = UIColor.white.cgColor
-        meetPaymentLabel.layer.cornerRadius = meetPaymentLabel.frame.height/2
-        
-        directPaymentLabel.layer.borderWidth = 1
-        directPaymentLabel.layer.borderColor = UIColor.white.cgColor
-        directPaymentLabel.layer.cornerRadius = directPaymentLabel.frame.height/2
-        
-        tabBarInit()
-        tabBarAppearanceInit()
-        tabBar.layer.borderWidth = 1
-        tabBar.layer.borderColor = UIColor(hexString: "EEEEEE").cgColor
-        
-        tabBar.selectedItem = tabBar.items![0] as UITabBarItem
-        tabBar(tabBar, didSelect: mukkabieTabItem)
-        
+//        self.view.addSubview(cartAlertView)
 
-        /*for menu in (modelStore?.menu)! {
-            orderByMenu[menu["foodNm"] as! String] = 0
-            if Int(menu["foodPrice"] as! String) != 0 {
-                priceByMenu[menu["foodNm"] as! String] = Int(menu["foodPrice"] as! String)
-            }
-        }
-        }*/
-
+        cartAlertView.center = self.view.center
+        
+//        cartAlertView.alpha = 0
+        cartAlertView.layer.masksToBounds = true
+        cartAlertView.layer.cornerRadius = 1
+        
+        NotificationCenter.default.addObserver(self, selector: #selector(getStore(_:)), name: NSNotification.Name(rawValue: "getStore"), object: nil)
         NotificationCenter.default.addObserver(self, selector: #selector(getOrderList(_:)), name: NSNotification.Name(rawValue: "getOrder"), object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(postOrder(_:)), name: NSNotification.Name(rawValue: "postOrder"), object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(changeTab(_:)), name: NSNotification.Name(rawValue: "changeTab"), object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(touchedSubTableView(_:)), name: NSNotification.Name(rawValue: "touchedSubTableView"), object: nil)
         
-        networkOrder.getOrderList(sellerId: (modelStore?.id)!)
+        self.networkStore.getStoreList(sellerId: storeId)
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        
+        indicatorView.showProgressView(view: self.view)
+
     }
     
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
     }
     
-    @IBAction func touchedShoppingCart(_ sender: Any) {
-        if priceByMenu.count > 0 {
-            let randomNo = arc4random_uniform(UInt32(priceByMenu.count))
-            networkOrder.postOrder(sellderId: (modelStore?.id)!, buyerId: "hjtech", price: Array(priceByMenu.values)[Int(randomNo)], content: [Array(priceByMenu.keys)[Int(randomNo)]])
+    func getStore(_ notification: Notification) {
+        guard let userInfo = notification.userInfo,
+            let storeInfo = userInfo["storeList"] as? [ModelStores] else { return }
+        self.modelStore = storeInfo[0]
+        
+        self.navigationItem.title = modelStore?.name
+        
+        mukkaebieVC?.modelStore = self.modelStore
+        menuRankVC?.modelStore = self.modelStore
+        infoVC?.introText = self.modelStore?.storeDesc
+        infoVC?.openHourText = self.modelStore?.openHour
+        infoVC?.telephoneText = self.modelStore?.telephone
+        infoVC?.nameText = self.modelStore?.name
+        
+        initMenuArray()
+        
+        self.networkOrder.getOrderList(sellerId: storeId)
+        tableView.reloadData()
+    }
+    
+    func initMenuArray() {
+        if (self.modelStore?.menu.count)! > 0 {
+            let menu = (self.modelStore?.menu)![0]
+            for (title, submenu) in menu {
+                let item = MenuViewModelItem(sectionTitle: title, rowCount: submenu.count, isCollapsed: false)
+                menuRankVC?.items.append(item)
+                var menu : [(key: String, value: String)] = []
+                for (name, price) in submenu {
+                    menu.append((key: name, value: price))
+                }
+                menuRankVC?.menus.append(menu)
+            }
+        }
+        initOrderByMenu()
+        initPriceByMenu()
+    }
+    
+    func initOrderByMenu() {
+        orderByMenu = [String:Int]()
+        if (self.modelStore?.menu.count)! > 0 {
+            let menu = (self.modelStore?.menu)![0]
+            for (_, submenu) in menu {
+                for (name, _) in submenu {
+                    orderByMenu[name] = 0
+                }
+            }
+        }
+    }
+    
+    func initPriceByMenu() {
+        priceByMenu = [String:Int]()
+        if (self.modelStore?.menu.count)! > 0 {
+            let menu = (self.modelStore?.menu)![0]
+            for (_, submenu) in menu {
+                for (name, price) in submenu {
+                    priceByMenu[name] = Int(price.substring(to: price.index(before: price.endIndex)).replacingOccurrences(of: ",", with: ""))
+                }
+            }
         }
     }
     
@@ -98,147 +160,239 @@ class StoreDetailViewController: UIViewController, UITabBarDelegate {
         guard let userInfo = notification.userInfo,
             let orderInfo = userInfo["orderList"] as? [ModelOrders] else { return }
         self.orderList = orderInfo
-
+        
+        getOrderByMenu()
+        getOrderByUser()
+        let indexPath = IndexPath(row: 0, section: 3)
+        tableView.reloadRows(at: [indexPath], with: .none)
+    }
+    
+    func getOrderByMenu() {
+        initOrderByMenu()
+        
         for order in self.orderList {
             for content in order.content {
                 orderByMenu[content]! += 1
             }
         }
-    
+        
         orderByMenuSorted = orderByMenu.sorted(by: { $0.1 > $1.1 })
         
         if orderByMenuSorted.count > 3 {
             var count = 0
-            for i in (3 ..< orderByMenuSorted.count-1).reversed() {
+            for i in (3 ..< orderByMenuSorted.count).reversed() {
                 count += orderByMenuSorted[i].value
                 orderByMenuSorted.removeLast()
             }
             orderByMenuSorted.append((key: "기타", value: count))
         }
+        
+        menuRankVC?.orderByMenuSorted = self.orderByMenuSorted
+        
+        if orderList.count == 0 {
+            menuRankVC?.noOrder = true
+        } else {
+            menuRankVC?.noOrder = false
+        }
+        if (menuRankVC?.pieChartView != nil) {
+            menuRankVC?.setSegment()
+        }
     }
     
-    func tabBar(_ tabBar: UITabBar, didSelect item: UITabBarItem) {
-        
-        for child in self.childViewControllers {
-            child.removeFromParentViewController()
+    func getOrderByUser() {
+        for order in self.orderList {
+            if orderByUser[order.buyerId] == nil {
+                orderByUser[order.buyerId] = 1
+            } else {
+                orderByUser[order.buyerId]! += 1
+            }
         }
         
-        for subview in tabSubView.subviews {
+        let orderByUserSorted = orderByUser.sorted(by: { $0.1 > $1.1 })
+        
+        for i in orderByUserSorted.count > 3 ? 0..<3 : 0..<orderByUserSorted.count {
+            self.orderByUserTop3.append(orderByUserSorted[i])
+        }
+    
+        mukkaebieVC?.orderByUserTop3 = self.orderByUserTop3
+    }
+    
+    func postOrder(_ notification: Notification) {
+        self.networkOrder.getOrderList(sellerId: storeId)
+    }
+    
+    func changeTab(_ notification: Notification) {
+        guard let userInfo = notification.userInfo,
+            let tabNumber = userInfo["tabNumber"] as? Int else { return }
+        
+        self.tabNumber = tabNumber
+        
+        let indexPath = IndexPath(row: 0, section: 3)
+        tableView.reloadRows(at: [indexPath], with: .none)
+    }
+    
+    func touchedSubTableView(_ notification: Notification) {
+        let indexPath = IndexPath(row: 0, section: 3)
+        tableView.reloadRows(at: [indexPath], with: .none)
+    }
+    
+    @IBAction func touchedShoppingCart(_ sender: Any) {
+        
+
+        self.view.addSubview(cartAlertView)
+        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now()+2) { 
+            self.cartAlertView.removeFromSuperview()
+        }
+    }
+    
+    @IBAction func touchedOrderCall(_ sender: Any) {
+        let url = NSURL(string: "tel://\(modelStore?.telephone as! String)")
+        UIApplication.shared.openURL(url as! URL)
+    }
+}
+
+
+
+extension StoreDetailViewController: UITableViewDataSource, UITableViewDelegate {
+    
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        if section == 0 {
+            let headerView = Bundle.main.loadNibNamed("statusTableViewCell", owner: self, options: nil)?.first as! statusTableViewCell
+            return headerView
+        }
+        
+        if section == 1 {
+            let logoView = Bundle.main.loadNibNamed("logoTableViewCell", owner: self, options: nil)?.first as! logoTableViewCell
+            if let url = URL(string: (modelStore?.imgURL)!) {
+                logoView.logoImage.af_setImage(withURL: url)
+            } else {
+                logoView.logoImage.image  = #imageLiteral(resourceName:"woowatech")
+            }
+            return logoView
+        }
+            
+        if section == 2 {
+            let rateView = Bundle.main.loadNibNamed("rateTableViewCell", owner: self, options: nil)?.first as! rateTableViewCell
+            rateView.ratingValue.text = String(describing: floor((modelStore?.ratingValue)!*10)/10)
+            return rateView
+        }
+        
+        if section == 3 {
+            let tapView = Bundle.main.loadNibNamed("segment", owner: self, options: nil)?.first as! segment
+            
+            tapView.segmentView.frame.size.width = self.view.frame.width
+            tapView.segmentView.updateView()
+            
+            return tapView.contentView
+        }
+        
+        return self.view
+    }
+    
+    
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        if section == 0 {
+            return 36
+        }
+        else if section == 1 {
+            return 155
+        }
+        else if section == 2 {
+            return 96
+        }
+        else if section == 3 {
+            return 40
+        }
+        return 0
+    }
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        if modelStore == nil {
+            return 0
+        }
+        return 4
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        if section == 0 {
+            return 0
+        }
+        else if section == 1 {
+            return 0
+        }
+        else if section == 2 {
+            return 0
+        }
+        else if section == 3 {
+            return 1
+        }
+        return 0
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
+        let cell = tableView.dequeueReusableCell(withIdentifier: "cell") as! TabSubviewTableViewCell
+        for subview in cell.tabSubview.subviews {
             subview.removeFromSuperview()
         }
-        
-        switch item.tag {
-            
+        switch tabNumber {
         case 0:
-            let storyboard = UIStoryboard(name: "MukkaebieRank", bundle: nil)
-            mukkaebieRankViewController = storyboard.instantiateViewController(withIdentifier: "MukkaebieRank") as? MukkaebieRankViewController
-            addChildViewController(mukkaebieRankViewController!)
-            
-            tabSubView.addSubview((mukkaebieRankViewController?.view)!)
-            mukkaebieRankViewController?.didMove(toParentViewController: self)
-            
-            adjustContentHeight(tabSubViewHeight: (mukkaebieRankViewController?.view.frame.size.height)!)
-            adjustContentConstraint()
-            break
-            
+            if orderByUserTop3.count == 0 {
+                let noMukkaebieCell = Bundle.main.loadNibNamed("noMukkaebieTableViewCell", owner: self, options: nil)?.first as! UITableViewCell
+                return noMukkaebieCell
+            } else {
+                cell.tabSubview.frame.size.height = (mukkaebieVC?.view.frame.height)!
+                cell.tabSubviewHeightConstraint.constant = (mukkaebieVC?.view.frame.height)!
+                cell.tabSubview.addSubview((mukkaebieVC?.view)!)
+            }
         case 1:
-            let storyboard = UIStoryboard(name: "MenuView", bundle: nil)
-            menuViewController = storyboard.instantiateViewController(withIdentifier: "Menu") as? MenuViewController
-            addChildViewController(menuViewController!)
-            
-            menuViewController?.orderByMenuSorted = orderByMenuSorted
-            
-            tabSubView.addSubview((menuViewController?.view)!)
-            menuViewController?.didMove(toParentViewController: self)
-                
-            let tabSubviewHeight = (menuViewController?.view.frame.size.height)! - (menuViewController?.menuTableView.frame.size.height)! + (menuViewController?.menuTableView.contentSize.height)!
-            adjustContentHeight(tabSubViewHeight: tabSubviewHeight)
-            adjustContentConstraint()
-            break
+            if modelStore?.menu.count == 0 {
+                let noMenuCell = Bundle.main.loadNibNamed("noMenuTableViewCell", owner: self, options: nil)?.first as! UITableViewCell
+                return noMenuCell
+            } else {
+                cell.tabSubview.frame.size.height = (menuRankVC?.view.frame.height)!
+                cell.tabSubviewHeightConstraint.constant = (menuRankVC?.view.frame.height)!
+                cell.tabSubview.addSubview((menuRankVC?.view)!)
+            }
         case 2:
-            break
+            cell.tabSubview.frame.size.height = (infoVC?.view.frame.height)!
+            cell.tabSubviewHeightConstraint.constant = (infoVC?.view.frame.height)!
+            cell.tabSubview.addSubview((infoVC?.view)!)
         case 3:
-            break
+            cell.tabSubview.frame.size.height = (reviewVC?.tableView.contentSize.height)!
+            cell.tabSubviewHeightConstraint.constant = (reviewVC?.tableView.contentSize.height)!
+            cell.tabSubview.addSubview((reviewVC?.view)!)
         default:
             break
         }
+        
+        return cell
     }
     
-    func tabBarInit() {
-        mukkabieTabItem.tag = 0
-        mukkabieTabItem.image = makeThumbnailFromText(text: "먹깨비", size: CGSize(width: tabBar.frame.width / CGFloat((tabBar.items?.count)!), height: tabBar.frame.height))
-        mukkabieTabItem.title = nil;
-        
-        menuTabItem.tag = 1
-        menuTabItem.image = makeThumbnailFromText(text: "메뉴", size: CGSize(width: tabBar.frame.width / CGFloat((tabBar.items?.count)!), height: tabBar.frame.height))
-        menuTabItem.title = nil;
-        
-        infoTabItem.tag = 2
-        infoTabItem.image = makeThumbnailFromText(text: "정보", size: CGSize(width: tabBar.frame.width / CGFloat((tabBar.items?.count)!), height: tabBar.frame.height))
-        infoTabItem.title = nil;
-        
-        reviewTabItem.tag = 3
-        reviewTabItem.image = makeThumbnailFromText(text: "리뷰", size: CGSize(width: tabBar.frame.width / CGFloat((tabBar.items?.count)!), height: tabBar.frame.height))
-        reviewTabItem.title = nil;
-    }
-    
-    func tabBarAppearanceInit() {
-        UITabBar.appearance().tintColor = UIColor(white: 136/255, alpha: 1)
-        UITabBar.appearance().barTintColor = UIColor.white
-        
-        //make image for selection - green rectagle
-        UIGraphicsBeginImageContextWithOptions(CGSize(width: tabBar.frame.width / CGFloat((tabBar.items?.count)!), height: tabBar.frame.height), false, 0)
-        UIColor.white.setFill()
-        UIRectFill(CGRect(x: 0, y: 0, width: tabBar.frame.width / CGFloat((tabBar.items?.count)!
-            ), height: tabBar.frame.height))
-        UIColor(red: 42/255, green: 193/255, blue: 188/255, alpha: 1).setFill()
-        UIRectFill(CGRect(x: 0, y: tabBar.frame.height - 5, width: tabBar.frame.width / CGFloat((tabBar.items?.count)!
-            ), height: 5))
-        let image = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-        UITabBar.appearance().selectionIndicatorImage = image
-    }
-    
-    func makeThumbnailFromText(text: String, size: CGSize) -> UIImage {
-        // some variables that control the size of the image we create, what font to use, etc.
-        
-        let imageSize = size
-        let fontSize: CGFloat = 15.0
-        let fontName = "AppleSDGothicNeo-Bold"
-        let font = UIFont(name: fontName, size: fontSize)!
-        
-        // set up the context and the font
-        UIGraphicsBeginImageContextWithOptions(imageSize, false, 0)
-        let attributes = [NSFontAttributeName: font]
-        
-        let x = (imageSize.width - text.size(attributes: attributes).width) / 2.0
-        let y = (imageSize.height - fontSize) / 2.0
-        text.draw(at: CGPoint(x: x, y: y), withAttributes: attributes)
-        
-        let image = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-        return image!
-    }
-    
-    func adjustContentHeight(tabSubViewHeight: CGFloat) {
-        tabSubView.frame.size.height = tabSubViewHeight
-        tabView.frame.size.height = tabBar.frame.size.height+tabSubView.frame.size.height
-        contentView.frame.size.height = tabView.frame.size.height+imageView.frame.size.height+scoreView.frame.size.height+uppperBarView.frame.size.height
-        scrollView.contentSize.height = contentView.frame.size.height
-    }
-    
-    func adjustContentConstraint() {
-        if let constraint = (tabSubView.constraints.filter{$0.firstAttribute == .height}.first) {
-            constraint.constant = tabSubView.frame.size.height
-            print(tabSubView.frame.size.height)
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        if modelStore == nil {
+            return 0
         }
-        if let constraint = (tabView.constraints.filter{$0.firstAttribute == .height}.first) {
-            constraint.constant = tabView.frame.size.height
+        switch tabNumber {
+        case 0:
+            if orderByUserTop3.count == 0 {
+                let noMukkaebieCell = Bundle.main.loadNibNamed("noMukkaebieTableViewCell", owner: self, options: nil)?.first as! UITableViewCell
+                return noMukkaebieCell.frame.height
+            }
+            return (mukkaebieVC?.view.frame.height)!
+        case 1:
+            if modelStore?.menu.count == 0 {
+                let noMenuCell = Bundle.main.loadNibNamed("noMenuTableViewCell", owner: self, options: nil)?.first as! UITableViewCell
+                return noMenuCell.frame.height
+            }
+            return (menuRankVC?.view.frame.height)!
+        case 2:
+            return (infoVC?.view.frame.height)!
+        case 3:
+            return (reviewVC?.tableView.contentSize.height)!
+        default:
+            break
         }
-        if let constraint = (contentView.constraints.filter{$0.firstAttribute == .height}.first) {
-            constraint.constant = contentView.frame.size.height
-        }
+        return 0
     }
 }
